requirements.txt
streamlit==1.32.0
python-dotenv==1.0.1
openai==1.12.0